Switch to using onChange for hooking location in tabs
